app/requirements.txt
fastapi==0.115.0
jinja2==3.1.6
uvicorn==0.32.1
python-multipart==0.0.20
sqlalchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10